+++ CHANGELOG.md
@@ -24,6 +24,7 @@ API Additions:
 - Arrows being shot while casting spell with `"mode": "ITEM_USE"`, or shot with `"type": "SHOOT_ARROW"` can perform impact actions of the spell, can have their custom projectile model
 - ItemConfig `attributes[].id` field now accepts projectile damage and combat roll related attributes. Third party attributes can be support via Mixin into `AttributeResolver`
 - Add `HealthImpacting` interface for status effects, to modify damage and healing taken
+- Add some shared status effect renderers: `OrbitingEffectRenderer`, `StunParticleSpawner`
 
 Other changes:
 - Update MixinExtras to 0.2.0